Base class for KeysTestCase
