+++ sdk/keyvault/azure-keyvault-keys/tests/test_crypto_client_async.py
@@ -22,7 +22,8 @@ import pytest
 
 from _shared.helpers_async  import get_completed_future
 from _shared.json_attribute_matcher import json_attribute_matcher
-from _test_case_async import KeysTestCase, suffixed_test_name
+from _test_case_async import KeysTestCase
+from _test_case_base import suffixed_test_name
 
 KeyVaultPreparer = functools.partial(
     PowerShellPreparer,
@@ -142,10 +143,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key = await self._create_ec_key(key_client, self.get_resource_name("eckey"), hardware_protected=is_hsm)
 
-        crypto_client = self.create_crypto_client(key.id)
+        crypto_client = self.create_crypto_client(key.id, is_async=True)
         await crypto_client._initialize()
         assert crypto_client.key_id == key.id
 
@@ -162,10 +163,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key = await self._create_rsa_key(key_client, self.get_resource_name("rsakey"), hardware_protected=is_hsm)
 
-        crypto_client = self.create_crypto_client(key.id)
+        crypto_client = self.create_crypto_client(key.id, is_async=True)
         await crypto_client._initialize()
         assert crypto_client.key_id == key.id
 
@@ -183,11 +184,11 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET)
+        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET, is_async=True)
         key_name = self.get_resource_name("keycrypt")
 
         imported_key = await self._import_test_key(key_client, key_name, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(imported_key.id)
+        crypto_client = self.create_crypto_client(imported_key.id, is_async=True)
 
         result = await crypto_client.encrypt(EncryptionAlgorithm.rsa_oaep, self.plaintext)
         self.assertEqual(result.key_id, imported_key.id)
@@ -204,7 +205,7 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET)
+        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET, is_async=True)
         key_name = self.get_resource_name("keysign")
 
         md = hashlib.sha256()
@@ -212,7 +213,7 @@ class CryptoClientTests(KeysTestCase):
         digest = md.digest()
 
         imported_key = await self._import_test_key(key_client, key_name, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(imported_key.id)
+        crypto_client = self.create_crypto_client(imported_key.id, is_async=True)
 
         result = await crypto_client.sign(SignatureAlgorithm.rs256, digest)
         self.assertEqual(result.key_id, imported_key.id)
@@ -229,12 +230,12 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET)
+        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET, is_async=True)
         key_name = self.get_resource_name("keywrap")
 
         created_key = await self._create_rsa_key(key_client, key_name, hardware_protected=is_hsm)
         self.assertIsNotNone(created_key)
-        crypto_client = self.create_crypto_client(created_key.id)
+        crypto_client = self.create_crypto_client(created_key.id, is_async=True)
 
         # Wrap a key with the created key, then unwrap it. The wrapped key's bytes should round-trip.
         key_bytes = self.plaintext
@@ -250,12 +251,12 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("symmetric-encrypt")
 
         imported_key = await self._import_symmetric_test_key(key_client, key_name)
         assert imported_key is not None
-        crypto_client = self.create_crypto_client(imported_key.id)
+        crypto_client = self.create_crypto_client(imported_key.id, is_async=True)
         # Use 256-bit AES algorithms for the 256-bit key
         symmetric_algorithms = [algo for algo in EncryptionAlgorithm if algo.startswith("A256")]
 
@@ -292,12 +293,12 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("symmetric-kw")
 
         imported_key = await self._import_symmetric_test_key(key_client, key_name)
         assert imported_key is not None
-        crypto_client = self.create_crypto_client(imported_key.id)
+        crypto_client = self.create_crypto_client(imported_key.id, is_async=True)
 
         result = await crypto_client.wrap_key(KeyWrapAlgorithm.aes_256, self.plaintext)
         assert result.key_id == imported_key.id
@@ -313,10 +314,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("encrypt-local")
         key = await self._create_rsa_key(key_client, key_name, size=4096, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(key)
+        crypto_client = self.create_crypto_client(key, is_async=True)
 
         rsa_encrypt_algorithms = [algo for algo in EncryptionAlgorithm if algo.startswith("RSA")]
         for encrypt_algorithm in rsa_encrypt_algorithms:
@@ -334,10 +335,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("encrypt-local")
         key = await self._create_rsa_key(key_client, key_name, size=4096, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(key)
+        crypto_client = self.create_crypto_client(key, is_async=True)
         local_client = CryptographyClient.from_jwk(key.key)
 
         rsa_encrypt_algorithms = [algo for algo in EncryptionAlgorithm if algo.startswith("RSA")]
@@ -356,10 +357,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("wrap-local")
         key = await self._create_rsa_key(key_client, key_name, size=4096, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(key)
+        crypto_client = self.create_crypto_client(key, is_async=True)
 
         for wrap_algorithm in (algo for algo in KeyWrapAlgorithm if algo.startswith("RSA")):
             result = await crypto_client.wrap_key(wrap_algorithm, self.plaintext)
@@ -376,10 +377,10 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         key_name = self.get_resource_name("wrap-local")
         key = await self._create_rsa_key(key_client, key_name, size=4096, hardware_protected=is_hsm)
-        crypto_client = self.create_crypto_client(key)
+        crypto_client = self.create_crypto_client(key, is_async=True)
         local_client = CryptographyClient.from_jwk(key.key)
 
         for wrap_algorithm in (algo for algo in KeyWrapAlgorithm if algo.startswith("RSA")):
@@ -397,11 +398,11 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         for size in (2048, 3072, 4096):
             key_name = self.get_resource_name("rsa-verify-{}".format(size))
             key = await self._create_rsa_key(key_client, key_name, size=size, hardware_protected=is_hsm)
-            crypto_client = self.create_crypto_client(key)
+            crypto_client = self.create_crypto_client(key, is_async=True)
             for signature_algorithm, hash_function in (
                 (SignatureAlgorithm.ps256, hashlib.sha256),
                 (SignatureAlgorithm.ps384, hashlib.sha384),
@@ -426,11 +427,11 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         for size in (2048, 3072, 4096):
             key_name = self.get_resource_name("rsa-verify-{}".format(size))
             key = await self._create_rsa_key(key_client, key_name, size=size, hardware_protected=is_hsm)
-            crypto_client = self.create_crypto_client(key)
+            crypto_client = self.create_crypto_client(key, is_async=True)
             local_client = CryptographyClient.from_jwk(key.key)
             for signature_algorithm, hash_function in (
                     (SignatureAlgorithm.ps256, hashlib.sha256),
@@ -456,7 +457,7 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         matrix = {
             KeyCurveName.p_256: (SignatureAlgorithm.es256, hashlib.sha256),
             KeyCurveName.p_256_k: (SignatureAlgorithm.es256_k, hashlib.sha256),
@@ -467,7 +468,7 @@ class CryptoClientTests(KeysTestCase):
         for curve, (signature_algorithm, hash_function) in sorted(matrix.items()):
             key_name = self.get_resource_name("ec-verify-{}".format(curve.value))
             key = await self._create_ec_key(key_client, key_name, curve=curve, hardware_protected=is_hsm)
-            crypto_client = self.create_crypto_client(key)
+            crypto_client = self.create_crypto_client(key, is_async=True)
 
             digest = hash_function(self.plaintext).digest()
 
@@ -485,7 +486,7 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url)
+        key_client = self.create_key_client(azure_keyvault_url, is_async=True)
         matrix = {
             KeyCurveName.p_256: (SignatureAlgorithm.es256, hashlib.sha256),
             KeyCurveName.p_256_k: (SignatureAlgorithm.es256_k, hashlib.sha256),
@@ -496,7 +497,7 @@ class CryptoClientTests(KeysTestCase):
         for curve, (signature_algorithm, hash_function) in sorted(matrix.items()):
             key_name = self.get_resource_name("ec-verify-{}".format(curve.value))
             key = await self._create_ec_key(key_client, key_name, curve=curve, hardware_protected=is_hsm)
-            crypto_client = self.create_crypto_client(key)
+            crypto_client = self.create_crypto_client(key, is_async=True)
             local_client = CryptographyClient.from_jwk(key.key)
 
             digest = hash_function(self.plaintext).digest()
@@ -515,9 +516,9 @@ class CryptoClientTests(KeysTestCase):
         self._skip_if_not_configured(is_hsm)
         azure_keyvault_url = self.managed_hsm_url if is_hsm else azure_keyvault_url
 
-        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET)
+        key_client = self.create_key_client(azure_keyvault_url, permissions=NO_GET, is_async=True)
         async def test_operations(key, expected_error_substrings, encrypt_algorithms, wrap_algorithms):
-            crypto_client = self.create_crypto_client(key)
+            crypto_client = self.create_crypto_client(key, is_async=True)
             for algorithm in encrypt_algorithms:
                 with pytest.raises(ValueError) as ex:
                     await crypto_client.encrypt(algorithm, self.plaintext)
